project/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
langchain==0.1.0
langgraph==0.0.17
openai==1.6.1
pydantic==2.5.3
websockets==12.0
pytest==7.4.3
httpx==0.26.0
python-multipart==0.0.6
jinja2==3.1.2
loguru==0.7.2
